requirements.txt
requests==2.31.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pandas>=2.1.3
numpy>=1.26.0
python-dotenv==1.0.0
pyyaml>=6.0.1
click==8.1.7
rich>=14.0.0,<15.0.0
flask==3.0.0
gunicorn==21.2.0
statsbombpy==1.13.0
beautifulsoup4>=4.11.1
lxml>=4.9.4
apscheduler==3.10.4
understat==0.1.14
aiohttp>=3.8.3
